lecture_transcription_corrector.py
import os
from pathlib import Path
from typing import Dict, List, Tuple

import assemblyai as aai
from dotenv import load_dotenv
import fitz  # PyMuPDF
import pdfplumber
import whisper

class LectureTranscriptionCorrector:
    def __init__(self, api_key: str = None):
        """
        講義書き起こし修正システムの初期化
        Args:
            api_key: AssemblyAI APIキー
        """
        # AssemblyAIの設定
        if api_key:
            aai.settings.api_key = api_key
        
        # Whisperモデルの読み込み（バックアップとして）
        self.whisper_model = whisper.load_model("base")
        
    def transcribe_audio(self, audio_file: str, use_whisper: bool = False) -> str:
        """
        音声ファイルから文字起こしを行う
        Args:
            audio_file: 音声ファイルのパス
            use_whisper: Whisperを使用するかどうか
        Returns:
            文字起こしテキスト
        """
        if use_whisper:
            # Whisperを使用した文字起こし
            result = self.whisper_model.transcribe(audio_file)
            return result["text"]
        else:
            # AssemblyAIを使用した文字起こし
            config = aai.TranscriptionConfig(
                language_code="ja",
            )
            transcriber = aai.Transcriber()
            transcript = transcriber.transcribe(audio_file, config)
            return transcript.text

    def extract_text_from_pdf(self, pdf_file: str) -> List[str]:
        """
        PDFから講義資料のテキストを抽出
        Args:
            pdf_file: PDFファイルのパス
        Returns:
            ページごとのテキストのリスト
        """
        texts = []
        
        # PyMuPDFを使用した基本的なテキスト抽出
        doc = fitz.open(pdf_file)
        for page in doc:
            text = page.get_text()
            texts.append(text)
        
        # より詳細な抽出が必要な場合はpdfplumberを使用
        with pdfplumber.open(pdf_file) as pdf:
            for page in pdf.pages:
                # テーブルや特殊なレイアウトの処理が必要な場合に使用
                pass
        
        return texts

    def correct_transcription(self, transcription: str, reference_texts: List[str]) -> str:
        """
        講義資料を参照して書き起こしを修正
        Args:
            transcription: 元の書き起こしテキスト
            reference_texts: 講義資料から抽出したテキストのリスト
        Returns:
            修正された書き起こしテキスト
        """
        # TODO: より高度な修正ロジックの実装
        # 1. 固有名詞の抽出と修正
        # 2. 文脈を考慮した修正
        # 3. 数字や専門用語の修正
        return transcription

    def evaluate_quality(self, original: str, corrected: str) -> Dict[str, float]:
        """
        修正の品質を評価
        Args:
            original: 元の書き起こしテキスト
            corrected: 修正後のテキスト
        Returns:
            品質評価の結果
        """
        # TODO: 評価指標の実装
        # - 文の自然さ
        # - 元の発言との一致度
        # - 固有名詞の正確性
        return {
            "naturalness": 0.0,
            "faithfulness": 0.0,
            "accuracy": 0.0
        }

def main():
    # 環境変数の読み込み
    load_dotenv()
    api_key = os.getenv("ASSEMBLYAI_API_KEY")
    
    # システムの初期化
    corrector = LectureTranscriptionCorrector(api_key=api_key)
    
    # 使用例
    audio_file = "path/to/lecture.mp3"
    pdf_file = "path/to/lecture.pdf"
    
    # 文字起こし
    transcription = corrector.transcribe_audio(audio_file)
    
    # 講義資料からのテキスト抽出
    reference_texts = corrector.extract_text_from_pdf(pdf_file)
    
    # 書き起こしの修正
    corrected_text = corrector.correct_transcription(transcription, reference_texts)
    
    # 品質評価
    quality_metrics = corrector.evaluate_quality(transcription, corrected_text)
    print(f"Quality Metrics: {quality_metrics}")

if __name__ == "__main__":
    main() 