requirements.txt
assemblyai
openai-whisper==20231117
PyMuPDF==1.23.26
pdfplumber==0.10.4
python-dotenv==1.0.1
moviepy==1.0.3 